Refactor SD card library and add thread-safety #10.
